+++ graphics/r_dtypes.h
@@ -30,11 +30,16 @@ namespace Fresa::Graphics
     //----------------------------------------
     struct WindowData {
         SDL_Window* window;
+        
         Vec2<> size;
         Vec2<> resolution;
         ui16 scale;
         ui16 refresh_rate = 0;
+        
         bool vsync;
+        
+        glm::mat4 view;
+        glm::mat4 proj;
     };
     //----------------------------------------