+++ CHANGELOG.md
@@ -10,9 +10,13 @@
 - compile time type name generator
 - component list documentation
 - state machine events can now be linked with new event system
+- render system list caller inside r_graphics.h
 
 **changed**
 - removed macros from state machine name generation and changed it to compile time names
+
+**up next**
+- projection and view matrices
 
 #### 0.3.3 - hey there! this is an event - _14 12 2021_